requirements.txt
# Core dependencies
anthropic>=0.25.0
pyyaml>=6.0
python-dateutil>=2.8.0

# Future dependencies (Phase 2+)
# requests>=2.31.0          # For price scraping
# beautifulsoup4>=4.12.0    # For web parsing  
# sqlite3                   # Built-in, for price history
# schedule>=1.2.0           # For timing optimization

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0